Split schema.ts into 3 files
