+++ src/models/admin.ts
@@ -1,0 +1,12 @@
+import { prop, getModelForClass } from "@typegoose/typegoose";
+
+export class Admin {
+  @prop()
+  public login?: string;
+  @prop()
+  public password?: string;
+  @prop()
+  public password_salt?: string;
+}
+
+export const AdminModel = getModelForClass(Admin);
