+++ src/models/project.ts
@@ -1,0 +1,26 @@
+import { prop, getModelForClass, Ref } from "@typegoose/typegoose";
+
+import { User } from "./user";
+
+export class Project {
+  @prop()
+  public title?: string;
+  @prop()
+  public short_description?: string;
+  @prop()
+  public description?: string;
+  @prop()
+  public status?: string;
+  @prop()
+  public cover_picture_url?: string;
+  @prop()
+  public published?: boolean;
+  @prop({ ref: "User" })
+  public owner?: Ref<User>;
+  @prop({ ref: "User" })
+  public membmers?: Ref<User>[];
+  @prop()
+  public likes?: number;
+}
+
+export const ProjectModel = getModelForClass(Project);
